#262 refactor code
