+++ base/src/main/java/io/github/sds100/keymapper/base/actions/ActionData.kt
@@ -1,5 +1,6 @@
 package io.github.sds100.keymapper.base.actions
 
+import io.github.sds100.keymapper.base.actions.talkback.TalkBackGestureType
 import io.github.sds100.keymapper.common.models.ShellExecutionMode
 import io.github.sds100.keymapper.common.utils.NodeInteractionType
 import io.github.sds100.keymapper.common.utils.Orientation
@@ -95,6 +96,7 @@ sealed class ActionData : Comparable<ActionData> {
                     { it.showVolumeUi },
                     { it.volumeStream },
                 )
+
                 else -> super.compareTo(other)
             }
         }
@@ -111,6 +113,7 @@ sealed class ActionData : Comparable<ActionData> {
                     { it.showVolumeUi },
                     { it.volumeStream },
                 )
+
                 else -> super.compareTo(other)
             }
         }
@@ -1040,6 +1043,7 @@ sealed class ActionData : Comparable<ActionData> {
                 { it.settingKey },
                 { it.value },
             )
+
             else -> super.compareTo(other)
         }
     }
