+++ base/src/main/java/io/github/sds100/keymapper/base/system/accessibility/BaseAccessibilityService.kt
@@ -29,7 +29,8 @@ import androidx.savedstate.SavedStateRegistryController
 import androidx.savedstate.SavedStateRegistryOwner
 import dagger.hilt.android.AndroidEntryPoint
 import io.github.sds100.keymapper.base.R
-import io.github.sds100.keymapper.base.actions.TalkBackGestureType
+import io.github.sds100.keymapper.base.actions.talkback.TalkBackGestureType
+import io.github.sds100.keymapper.base.actions.talkback.TalkbackGesturePerformer
 import io.github.sds100.keymapper.base.input.InputEventDetectionSource
 import io.github.sds100.keymapper.common.utils.InputEventAction
 import io.github.sds100.keymapper.common.utils.KMError
@@ -562,6 +563,7 @@ abstract class BaseAccessibilityService :
         return imeWindow != null && imeWindow.root?.isVisibleToUser == true
     }
 
+    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
     override fun injectText(text: String) {
         inputMethod?.currentInputConnection?.commitText(
             text,
@@ -571,314 +573,12 @@ abstract class BaseAccessibilityService :
         )
     }
 
+    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
     override fun performImeAction() {
         inputMethod?.currentInputConnection?.performEditorAction(EditorInfo.IME_ACTION_UNSPECIFIED)
     }
 
     override fun performTalkBackGesture(gesture: TalkBackGestureType): KMResult<*> {
-        val dm = resources.displayMetrics
-        val cx = dm.widthPixels / 2f
-        val cy = dm.heightPixels / 2f
-        // Use 40% of the smaller screen dimension as swipe length
-        val swipeLen = minOf(dm.widthPixels, dm.heightPixels) * 0.4f
-        // Finger spacing for multi-finger gestures (pixels)
-        val fingerSpacing = dm.density * 40f
-
-        val gestureBuilder = GestureDescription.Builder()
-
-        when (gesture) {
-            TalkBackGestureType.SWIPE_UP ->
-                gestureBuilder.addStroke(buildSwipe(cx, cy, cx, cy - swipeLen, 200))
-
-            TalkBackGestureType.SWIPE_DOWN ->
-                gestureBuilder.addStroke(buildSwipe(cx, cy, cx, cy + swipeLen, 200))
-
-            TalkBackGestureType.SWIPE_LEFT ->
-                gestureBuilder.addStroke(buildSwipe(cx, cy, cx - swipeLen, cy, 200))
-
-            TalkBackGestureType.SWIPE_RIGHT ->
-                gestureBuilder.addStroke(buildSwipe(cx, cy, cx + swipeLen, cy, 200))
-
-            TalkBackGestureType.SWIPE_UP_THEN_DOWN -> {
-                val s1 = buildChainedSwipe(cx, cy, cx, cy - swipeLen, 150, willContinue = true)
-                val s2 = s1.continueStroke(buildPath(cx, cy - swipeLen, cx, cy), 150, 150, false)
-                gestureBuilder.addStroke(s1).addStroke(s2)
-            }
-
-            TalkBackGestureType.SWIPE_DOWN_THEN_UP -> {
-                val s1 = buildChainedSwipe(cx, cy, cx, cy + swipeLen, 150, willContinue = true)
-                val s2 = s1.continueStroke(buildPath(cx, cy + swipeLen, cx, cy), 150, 150, false)
-                gestureBuilder.addStroke(s1).addStroke(s2)
-            }
-
-            TalkBackGestureType.SWIPE_LEFT_THEN_RIGHT -> {
-                val s1 = buildChainedSwipe(cx, cy, cx - swipeLen, cy, 150, willContinue = true)
-                val s2 = s1.continueStroke(buildPath(cx - swipeLen, cy, cx, cy), 150, 150, false)
-                gestureBuilder.addStroke(s1).addStroke(s2)
-            }
-
-            TalkBackGestureType.SWIPE_RIGHT_THEN_LEFT -> {
-                val s1 = buildChainedSwipe(cx, cy, cx + swipeLen, cy, 150, willContinue = true)
-                val s2 = s1.continueStroke(buildPath(cx + swipeLen, cy, cx, cy), 150, 150, false)
-                gestureBuilder.addStroke(s1).addStroke(s2)
-            }
-
-            TalkBackGestureType.SWIPE_RIGHT_THEN_UP -> {
-                val s1 = buildChainedSwipe(cx, cy, cx + swipeLen, cy, 150, willContinue = true)
-                val s2 = s1.continueStroke(
-                    buildPath(cx + swipeLen, cy, cx, cy - swipeLen),
-                    150,
-                    150,
-                    false,
-                )
-                gestureBuilder.addStroke(s1).addStroke(s2)
-            }
-
-            TalkBackGestureType.TWO_FINGER_TAP ->
-                addMultiFingerTaps(
-                    gestureBuilder,
-                    cx,
-                    cy,
-                    fingerSpacing,
-                    fingerCount = 2,
-                    tapCount = 1,
-                    holdDuration = 50,
-                )
-
-            TalkBackGestureType.TWO_FINGER_DOUBLE_TAP_HOLD ->
-                addMultiFingerDoubleTapHold(gestureBuilder, cx, cy, fingerSpacing, fingerCount = 2)
-
-            TalkBackGestureType.TWO_FINGER_TRIPLE_TAP ->
-                addMultiFingerTaps(
-                    gestureBuilder,
-                    cx,
-                    cy,
-                    fingerSpacing,
-                    fingerCount = 2,
-                    tapCount = 3,
-                    holdDuration = 50,
-                )
-
-            TalkBackGestureType.TWO_FINGER_TRIPLE_TAP_HOLD ->
-                addMultiFingerTripleTapHold(gestureBuilder, cx, cy, fingerSpacing, fingerCount = 2)
-
-            TalkBackGestureType.THREE_FINGER_TAP ->
-                addMultiFingerTaps(
-                    gestureBuilder,
-                    cx,
-                    cy,
-                    fingerSpacing,
-                    fingerCount = 3,
-                    tapCount = 1,
-                    holdDuration = 50,
-                )
-
-            TalkBackGestureType.THREE_FINGER_TAP_HOLD ->
-                addMultiFingerTaps(
-                    gestureBuilder,
-                    cx,
-                    cy,
-                    fingerSpacing,
-                    fingerCount = 3,
-                    tapCount = 1,
-                    holdDuration = 600,
-                )
-
-            TalkBackGestureType.THREE_FINGER_TRIPLE_TAP_HOLD ->
-                addMultiFingerTripleTapHold(gestureBuilder, cx, cy, fingerSpacing, fingerCount = 3)
-
-            TalkBackGestureType.THREE_FINGER_SWIPE_UP ->
-                addMultiFingerSwipe(
-                    gestureBuilder,
-                    cx,
-                    cy,
-                    cx,
-                    cy - swipeLen,
-                    fingerSpacing,
-                    fingerCount = 3,
-                )
-
-            TalkBackGestureType.THREE_FINGER_SWIPE_DOWN ->
-                addMultiFingerSwipe(
-                    gestureBuilder,
-                    cx,
-                    cy,
-                    cx,
-                    cy + swipeLen,
-                    fingerSpacing,
-                    fingerCount = 3,
-                )
-
-            TalkBackGestureType.FOUR_FINGER_TAP ->
-                addMultiFingerTaps(
-                    gestureBuilder,
-                    cx,
-                    cy,
-                    fingerSpacing,
-                    fingerCount = 4,
-                    tapCount = 1,
-                    holdDuration = 50,
-                )
-
-            TalkBackGestureType.FOUR_FINGER_DOUBLE_TAP ->
-                addMultiFingerTaps(
-                    gestureBuilder,
-                    cx,
-                    cy,
-                    fingerSpacing,
-                    fingerCount = 4,
-                    tapCount = 2,
-                    holdDuration = 50,
-                )
-
-            TalkBackGestureType.FOUR_FINGER_SWIPE_UP ->
-                addMultiFingerSwipe(
-                    gestureBuilder,
-                    cx,
-                    cy,
-                    cx,
-                    cy - swipeLen,
-                    fingerSpacing,
-                    fingerCount = 4,
-                )
-
-            TalkBackGestureType.FOUR_FINGER_SWIPE_DOWN ->
-                addMultiFingerSwipe(
-                    gestureBuilder,
-                    cx,
-                    cy,
-                    cx,
-                    cy + swipeLen,
-                    fingerSpacing,
-                    fingerCount = 4,
-                )
-
-            TalkBackGestureType.FOUR_FINGER_SWIPE_LEFT ->
-                addMultiFingerSwipe(
-                    gestureBuilder,
-                    cx,
-                    cy,
-                    cx - swipeLen,
-                    cy,
-                    fingerSpacing,
-                    fingerCount = 4,
-                )
-
-            TalkBackGestureType.FOUR_FINGER_SWIPE_RIGHT ->
-                addMultiFingerSwipe(
-                    gestureBuilder,
-                    cx,
-                    cy,
-                    cx + swipeLen,
-                    cy,
-                    fingerSpacing,
-                    fingerCount = 4,
-                )
-        }
-
-        val success = dispatchGesture(gestureBuilder.build(), null, gestureHandler)
-        return if (success) Success(Unit) else KMError.FailedToDispatchGesture
-    }
-
-    private fun buildPath(x1: Float, y1: Float, x2: Float, y2: Float): Path = Path().apply {
-        moveTo(x1, y1)
-        lineTo(x2, y2)
-    }
-
-    private fun buildSwipe(
-        x1: Float,
-        y1: Float,
-        x2: Float,
-        y2: Float,
-        duration: Long,
-    ): StrokeDescription = StrokeDescription(buildPath(x1, y1, x2, y2), 0, duration)
-
-    private fun buildChainedSwipe(
-        x1: Float,
-        y1: Float,
-        x2: Float,
-        y2: Float,
-        duration: Long,
-        willContinue: Boolean,
-    ): StrokeDescription = StrokeDescription(buildPath(x1, y1, x2, y2), 0, duration, willContinue)
-
-    private fun fingerPositions(
-        cx: Float,
-        cy: Float,
-        spacing: Float,
-        count: Int,
-    ): List<Pair<Float, Float>> {
-        val total = (count - 1) * spacing
-        val start = cx - total / 2f
-        return (0 until count).map { i -> Pair(start + i * spacing, cy) }
-    }
-
-    private fun addMultiFingerTaps(
-        builder: GestureDescription.Builder,
-        cx: Float,
-        cy: Float,
-        spacing: Float,
-        fingerCount: Int,
-        tapCount: Int,
-        holdDuration: Long,
-    ) {
-        val positions = fingerPositions(cx, cy, spacing, fingerCount)
-        val tapInterval = 200L
-        for (tap in 0 until tapCount) {
-            val startTime = tap * tapInterval
-            for ((x, y) in positions) {
-                val path = Path().apply { moveTo(x, y) }
-                builder.addStroke(StrokeDescription(path, startTime, holdDuration))
-            }
-        }
-    }
-
-    private fun addMultiFingerDoubleTapHold(
-        builder: GestureDescription.Builder,
-        cx: Float,
-        cy: Float,
-        spacing: Float,
-        fingerCount: Int,
-    ) {
-        val positions = fingerPositions(cx, cy, spacing, fingerCount)
-        for ((x, y) in positions) {
-            val path = Path().apply { moveTo(x, y) }
-            builder.addStroke(StrokeDescription(path, 0, 50))
-            builder.addStroke(StrokeDescription(path, 200, 600))
-        }
-    }
-
-    private fun addMultiFingerTripleTapHold(
-        builder: GestureDescription.Builder,
-        cx: Float,
-        cy: Float,
-        spacing: Float,
-        fingerCount: Int,
-    ) {
-        val positions = fingerPositions(cx, cy, spacing, fingerCount)
-        for ((x, y) in positions) {
-            val path = Path().apply { moveTo(x, y) }
-            builder.addStroke(StrokeDescription(path, 0, 50))
-            builder.addStroke(StrokeDescription(path, 200, 50))
-            builder.addStroke(StrokeDescription(path, 400, 600))
-        }
-    }
-
-    private fun addMultiFingerSwipe(
-        builder: GestureDescription.Builder,
-        xStart: Float,
-        yStart: Float,
-        xEnd: Float,
-        yEnd: Float,
-        spacing: Float,
-        fingerCount: Int,
-    ) {
-        val startPositions = fingerPositions(xStart, yStart, spacing, fingerCount)
-        val endPositions = fingerPositions(xEnd, yEnd, spacing, fingerCount)
-        for (i in 0 until fingerCount) {
-            val (sx, sy) = startPositions[i]
-            val (ex, ey) = endPositions[i]
-            builder.addStroke(StrokeDescription(buildPath(sx, sy, ex, ey), 0, 200))
-        }
+        return TalkbackGesturePerformer.performTalkBackGesture(this, gesture, gestureHandler)
     }
 }
